recognizes requirements and interfaces on node templates
plus tidy up

+++ src/main/javac/org/apache/brooklyn/tosca/a4c/Alien4CloudToscaPlatform.java
@@ -1,10 +1,10 @@
 package org.apache.brooklyn.tosca.a4c;
 
 import java.io.Closeable;
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.file.FileSystem;
-import java.nio.file.FileSystems;
+import java.io.InputStream;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
@@ -12,10 +12,13 @@ import java.util.Properties;
 
 import org.apache.brooklyn.util.collections.MutableList;
 import org.apache.brooklyn.util.core.ResourceUtils;
-import org.apache.brooklyn.util.core.file.ArchiveUtils;
+import org.apache.brooklyn.util.core.file.ArchiveBuilder;
 import org.apache.brooklyn.util.exceptions.Exceptions;
+import org.apache.brooklyn.util.os.Os;
 import org.apache.brooklyn.util.text.Identifiers;
+import org.apache.brooklyn.util.text.Strings;
 import org.apache.brooklyn.util.time.Duration;
+import org.apache.commons.io.FileUtils;
 import org.elasticsearch.common.collect.Iterables;
 import org.elasticsearch.common.io.Streams;
 import org.slf4j.Logger;
@@ -36,7 +39,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import com.google.common.base.Stopwatch;
 
 import alien4cloud.csar.services.CsarService;
-import alien4cloud.git.RepositoryManager;
 import alien4cloud.model.components.Csar;
 import alien4cloud.security.model.Role;
 import alien4cloud.tosca.ArchiveUploadService;
@@ -50,14 +52,17 @@ public class Alien4CloudToscaPlatform implements Closeable {
     
     private static final Logger log = LoggerFactory.getLogger(Alien4CloudToscaPlatform.class);
     
+    public static final String TOSCA_NORMATIVE_TYPES_LOCAL_URL = "classpath://org/apache/brooklyn/tosca/a4c/tosca-normative-types.zip";
+    public static final String TOSCA_NORMATIVE_TYPES_GITHUB_URL = "https://github.com/alien4cloud/tosca-normative-types/archive/master.zip";
+
     public static Alien4CloudToscaPlatform newInstance(String ...args) throws Exception {
         log.info("Loading Alien4Cloud platform...");
+        // TODO support pre-existing ES instance
         // TODO if ES cannot find its config file, it will hang waiting for peers; should warn if does not complete in 1m
         try {
             Stopwatch s = Stopwatch.createStarted();
-            configure();
             AnnotationConfigApplicationContext ctx = new AnnotationConfigApplicationContext();
-            // TODO messy, manually loading the properties, but it works
+            // messy, manually loading the properties, but it works
             ctx.getEnvironment().getPropertySources().addFirst(new PropertiesPropertySource("user", 
                 A4CSpringConfig.alienConfig(ctx)));
             ctx.register(A4CSpringConfig.class);
@@ -77,61 +82,41 @@ public class Alien4CloudToscaPlatform implements Closeable {
             MutableList.of(new SimpleGrantedAuthority(Role.ADMIN.name())) ));
     }
 
-    public static void loadNormativeTypes(Alien4CloudToscaPlatform platform) throws Exception {
-        
-        // load normative types
-        String localName = "tosca-normative-types";
-        RepositoryManager repositoryManager = new RepositoryManager();
+    public void loadNormativeTypes() throws Exception {
+        Path zpc;
+        if (new ResourceUtils(this).doesUrlExist(TOSCA_NORMATIVE_TYPES_LOCAL_URL)) {
+            zpc = Paths.get(tmpRoot.toString(), "tosca-normative-types_"+Identifiers.makeRandomId(6)+".tgz");
+            Streams.copy(new ResourceUtils(this).getResourceFromUrl(TOSCA_NORMATIVE_TYPES_LOCAL_URL),
+                new FileOutputStream(zpc.toString()));
 
-        // TODO use zip
-        Path artifactsDirectory = Paths.get("../target/it-artifacts");
-        Path zpb = artifactsDirectory.resolve(Paths.get("tosca-normative-types."+Identifiers.makeRandomId(6)));
-        Path zpo = Paths.get(zpb.toString()+".orig.zip");
-        Streams.copy(new ResourceUtils(Alien4CloudToscaPlatform.class).getResourceFromUrl("https://github.com/alien4cloud/tosca-normative-types/archive/master.zip"),
-            new FileOutputStream(zpo.toString()) );
-        Path zpe = Paths.get(zpb.toString()+"_expanded");
-        FileUtil.unzip(zpo, zpe);
-        String zipRootDir = Iterables.getOnlyElement(Arrays.asList(zpe.toFile().list()));
-        Path zpc = Paths.get(zpb.toString()+".csar.zip");
-        FileUtil.zip(zpe.resolve(zipRootDir), zpc);
+        } else {
+            log.debug("Loading TOSCA normative types from GitHub");
+            // mainly kept for reference, in case we want to load other items from GitHub
+            Path artifactsDirectory = Paths.get(tmpRoot.toString(), "tosca-normative-types");
+            Path zpb = artifactsDirectory.resolve(Paths.get("tosca-normative-types."+Identifiers.makeRandomId(6)));
+            Path zpo = Paths.get(zpb.toString()+".orig.zip");
+            Streams.copy(new ResourceUtils(Alien4CloudToscaPlatform.class).getResourceFromUrl(TOSCA_NORMATIVE_TYPES_GITHUB_URL),
+                new FileOutputStream(zpo.toString()) );
+            Path zpe = Paths.get(zpb.toString()+"_expanded");
+            FileUtil.unzip(zpo, zpe);
+            String zipRootDir = Iterables.getOnlyElement(Arrays.asList(zpe.toFile().list()));
+            zpc = Paths.get(zpb.toString()+".csar.zip");
+            FileUtil.zip(zpe.resolve(zipRootDir), zpc);
+        }
         
-//        repositoryManager.cloneOrCheckout(artifactsDirectory, "https://github.com/alien4cloud/tosca-normative-types.git", "master", localName);
-//
-//        Path normativeTypesPath = artifactsDirectory.resolve(localName);
-//        Path normativeTypesZipPath = artifactsDirectory.resolve(localName + ".zip2");
-//        // Update zip
-//        FileUtil.zip(normativeTypesPath, normativeTypesZipPath);
-        
-////
-////        // Path normativeTypesZipPath = Paths.get("../target/it-artifacts/zipped/apache-lb-types-0.1.csar");
-////        ArchiveParser archiveParser = platform.getBean(ArchiveParser.class);
-////        ParsingResult<ArchiveRoot> normative = archiveParser.parse(normativeTypesZipPath);
-////        if (!normative.getContext().getParsingErrors().isEmpty()) {
-////            System.out.println("ERRORS:\n  "+Strings.join(tp.getContext().getParsingErrors(), "\n  "));
-////        }
-////        platform.getCsarService().save(normative.getResult().getArchive());
-////        // (save other items)
-//        
-////        String x = "/Users/alex/dev/gits/alien4cloud/alien4cloud/alien4cloud-core/src/test/resources/alien/paas/plan/csars/tosca-base-types-1.0";
-        
-        ParsingResult<Csar> normative = platform.getBean(ArchiveUploadService.class).upload(zpc);
-        
+        ParsingResult<Csar> normative = getBean(ArchiveUploadService.class).upload(zpc);
+
         if (ArchiveUploadService.hasError(normative, ParsingErrorLevel.ERROR))
-            throw new IllegalArgumentException("Errors parsing tosca normative types");
-    }
-
-    public static void configure() {
-        // TODO desired? A4C code has this, but seems redundant.
-//      if (Strings.isBlank(System.getProperty("spring.config.name"))) {
-//          System.setProperty("security.basic.enabled", "false");
-//          System.setProperty("spring.config.name", AlienYamlPropertiesFactoryBeanFactory.ALIEN_CONFIGURATION);
-//      }
+            throw new IllegalArgumentException("Errors parsing tosca normative types:\n"+Strings.join(normative.getContext().getParsingErrors(), "\n  "));
     }
 
     private ConfigurableApplicationContext ctx;   
-
+    File tmpRoot;
+    
     private Alien4CloudToscaPlatform(ConfigurableApplicationContext ctx) {
         this.ctx = ctx;
+        tmpRoot = Os.newTempDir("brooklyn-a4c");
+        Os.deleteOnExitRecursively(tmpRoot);
     }
 
     public synchronized void close() {
@@ -140,6 +125,7 @@ public class Alien4CloudToscaPlatform implements Closeable {
             ctx.close();
             ctx = null;
         }
+        Os.deleteRecursively(tmpRoot);
     }
     
     public <T> T getBean(Class<T> type) {
@@ -151,6 +137,9 @@ public class Alien4CloudToscaPlatform implements Closeable {
     }
     public CsarService getCsarService() {
         return getBean(CsarService.class);
+    }
+    public ArchiveUploadService getArchiveUploadService() {
+        return getBean(ArchiveUploadService.class);
     }
 
     
@@ -165,6 +154,36 @@ public class Alien4CloudToscaPlatform implements Closeable {
         @Bean(name={"alienconfig", "elasticsearchConfig"}) 
         public static Properties alienConfig(ResourceLoader resourceLoader) throws IOException {
             return AlienYamlPropertiesFactoryBeanFactory.get(resourceLoader).getObject();
+        }
+    }
+
+    public ParsingResult<Csar> uploadSingleYaml(InputStream resourceFromUrl, String nameO) {
+        try {
+            String name = Strings.makeValidFilename(nameO);
+            File tmpBase = new File(tmpRoot, name+"_"+Identifiers.makeRandomId(6));
+            File tmpExpanded = new File(tmpBase, name+"_"+Identifiers.makeRandomId(6));
+            File tmpTarget = new File(tmpExpanded.toString()+".csar.zip");
+            tmpExpanded.mkdir();
+
+            FileUtils.copyInputStreamToFile(resourceFromUrl, new File(tmpExpanded, name));
+            ArchiveBuilder.archive(tmpTarget.toString()).addDirContentsAt(tmpExpanded, "").create();
+
+            ParsingResult<Csar> result = getArchiveUploadService().upload(Paths.get(tmpTarget.toString()));
+            
+            Os.deleteRecursively(tmpBase);
+            
+            if (ArchiveUploadService.hasError(result, null)) {
+                log.debug("A4C parse notes for "+nameO+":\n  "+Strings.join(result.getContext().getParsingErrors(), "\n  "));
+            }
+            if (ArchiveUploadService.hasError(result, ParsingErrorLevel.ERROR)) {
+                throw new IllegalArgumentException("Errors parsing "+name+"; archive will not be installed:\n  "
+                    +Strings.join(result.getContext().getParsingErrors(), "\n  "));
+            }
+            
+            return result;
+
+        } catch (Exception e) {
+            throw Exceptions.propagate(e);
         }
     }
